candidate/requirements.txt
pytest==7.4.3
requests==2.31.0
psycopg2-binary==2.9.9
playwright==1.40.0
pytest-playwright==0.4.3
